Rename packages and commands to neutral names
Step 2 of generalizing the toolkit (GENERALIZATION_PLAN.md). Drop "radx" from
the Python package, distribution, and command names so the tools read as a
general data-dictionary toolkit:

- package radx_dd_converter -> dd_converter (git-tracked rename; history kept)
- dist radx-dd-converter -> dd-converter
- commands radx-dd-to-linkml -> dd-to-linkml, linkml-to-radx-dd -> linkml-to-dd
  (printer's dd-print unchanged)
- update every import across converter + printer + tests, the entry points,
  package-find/package-data, the printer's converter dependency, argparse prog
  names, and the generated-schema header comment; regenerate the examples.

The repository name is intentionally unchanged, so install URLs still contain
"radx" while the installed commands are neutral. Prose/spec generalization is a
separate follow-up (step 3). 152 tests pass (142 converter + 10 printer);
ruff-clean; examples lint clean.
