+++ converter/examples/rad.yaml
@@ -1,4 +1,4 @@
-# LinkML schema generated from a RADx data dictionary by radx-dd-to-linkml.
+# LinkML schema generated from a data dictionary by dd-to-linkml.
 #
 # The single class below (the tree_root) describes a target datafile: each of
 # its slots corresponds to one field (column) of that datafile, in order. The
