+++ converter/pyproject.toml
@@ -3,9 +3,9 @@ requires = ["setuptools>=61"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "radx-dd-converter"
+name = "dd-converter"
 version = "0.0.1"
-description = "Convert a RADx data dictionary CSV into a LinkML schema"
+description = "Convert a data dictionary CSV into a LinkML schema (and back)"
 readme = "README.md"
 requires-python = ">=3.9"
 license = { text = "BSD-2-Clause" }
@@ -18,15 +18,15 @@ dependencies = [
 test = ["pytest>=7"]
 
 [project.scripts]
-radx-dd-to-linkml = "radx_dd_converter.cli:main"
-linkml-to-radx-dd = "radx_dd_converter.reverse:main"
+dd-to-linkml = "dd_converter.cli:main"
+linkml-to-dd = "dd_converter.reverse:main"
 
 [tool.setuptools.packages.find]
 where = ["."]
-include = ["radx_dd_converter*"]
+include = ["dd_converter*"]
 
 [tool.setuptools.package-data]
-radx_dd_converter = ["grammar/*.lark"]
+dd_converter = ["grammar/*.lark"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
