+++ converter/tests/test_cli.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 import yaml
 
-from radx_dd_converter.cli import _class_from_name, _name_from_filename, main
+from dd_converter.cli import _class_from_name, _name_from_filename, main
 
 FIXTURE = Path(__file__).parent / "fixtures" / "sample.csv"
